Fix escape_quotes() edge case
There is an edge case when the quote escaping function does not work as
expected: When the parameters to be escaped match an argument of the
actual `echo` binary.

For example, when you try to escape '--version', it gets expanded to
"/bin/echo --version", which outputs the version of the echo utility
instead of the string '--version'.

This can be avoided by forcing `echo` to not parse any arguments via the
`POSIXLY_CORRECT` env variable.

This affects the `echo` version of the GNU coreutils package.

Reference point:
http://git.savannah.gnu.org/cgit/coreutils.git/tree/src/echo.c#n111